+++ allauth/headless/mfa/views.py
@@ -9,6 +9,7 @@ from allauth.headless.mfa.inputs import (
     AuthenticateInput,
     GenerateRecoveryCodesInput,
 )
+from allauth.mfa.adapter import DefaultMFAAdapter, get_adapter
 from allauth.mfa.models import Authenticator
 from allauth.mfa.recovery_codes.internal import flows as recovery_codes_flows
 from allauth.mfa.stages import AuthenticateStage
@@ -44,14 +45,17 @@ class AuthenticatorsView(AuthenticatedAPIView):
         return response.AuthenticatorsResponse(request, authenticators)
 
 
+# TODO: Enforce reauthentication
 class ManageTOTPView(AuthenticatedAPIView):
     input_class = {"POST": ActivateTOTPInput}
 
     def get(self, request, *args, **kwargs):
         authenticator = self._get_authenticator()
         if not authenticator:
+            adapter: DefaultMFAAdapter = get_adapter()
             secret = totp_auth.get_totp_secret(regenerate=True)
-            return response.TOTPNotFoundResponse(request, secret)
+            totp_url: str = adapter.build_totp_url(request.user, secret)
+            return response.TOTPNotFoundResponse(request, secret, totp_url)
         return response.TOTPResponse(request, authenticator)
 
     def _get_authenticator(self):